microservices/web/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-property==0.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jwcrypto==1.5.6
ldap3==2.9.1
MarkupSafe==3.0.2
packaging==24.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
python-dotenv==1.0.1
python-keycloak==4.7.2
redis==5.2.0
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
